pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.3",
    "python-dotenv>=1.1.1",
    "langchain-google-genai>=0.1.6",
]
